cnclip_api_service/requirements.txt
fastapi  
uvicorn  
torch  
Pillow  
cn_clip